+++ ydb/core/kqp/proxy_service/kqp_script_executions_ut.cpp
@@ -248,7 +248,7 @@ struct TScriptExecutionsYdbSetup {
         NYdb::NTable::TTableClient client(*YdbDriver, NYdb::NTable::TClientSettings().AuthToken(""));
         const auto sessionResult = client.CreateSession().ExtractValueSync();
         UNIT_ASSERT_VALUES_EQUAL_C(sessionResult.GetStatus(), NYdb::EStatus::SUCCESS, sessionResult.GetIssues().ToOneLineString());
-        const auto result = sessionResult.GetSession().DescribeTable(JoinPath({"dc-1", JoinPath(pathComponents)})).ExtractValueSync();
+        const auto result = sessionResult.GetSession().DescribeTable("/" + JoinPath({"dc-1", JoinPath(pathComponents)})).ExtractValueSync();
         UNIT_ASSERT_VALUES_EQUAL_C(result.GetStatus(), NYdb::EStatus::SUCCESS, result.GetIssues().ToOneLineString());
 
         const auto& tableDesc = result.GetTableDescription();