+++ ydb/public/lib/ydb_cli/common/normalize_path.cpp
@@ -45,21 +45,15 @@ namespace NConsoleClient {
     }
 
     void AdjustPath(TString& path, const TClientCommand::TConfig& config) {
-        if (path.StartsWith('/')) {
-            if (!path.StartsWith(config.Database)) {
-                throw TMisuseException() << "Provided path \"" << path << "\" starts with '/'. "
-                    << "That means you are using an absolute path that should start with the path "
-                    << "to your database \"" << config.Database << "\", but it doesn't. " << Endl
-                    << "Please, provide full path starting from the domain root "
-                    << "(example: \"/domain/my_base/dir1/table1\"). " << Endl
-                    << "Or consider using relative path from your database (example: \"dir1/table1\").";
-            }
-        } else {
-            // allow relative path
-            path = (config.Path ? config.Path : config.Database) + '/' + path;
+        if (!path.StartsWith('/') && config.Path) {
+            path = config.Path + '/' + path;
         }
 
+        // The server knows the cluster root and resolves database-relative paths.
         path = NormalizePath(path);
+        while (path.StartsWith("./")) {
+            path = path.substr(2);
+        }
     }
 
 }